Add files via upload
In this project we have utilized ML models for classification to predict if a borrower will fail to pay off the loan or not. We have used a vast dataset composed of past loans and visualized the data to recognize its scope and fallacies. Then we cleaned the data and identified the dominant features. After feature selection, we have executed model fitting. We have utilized k Nearest Neighbors, Naive Bayes, Logistic Regression, Support Vector Machine and Decision Tree models. Further we have implemented hyperparameter tuning to achieve the maximum accuracy on the test data.